+++ packages/orq-rc/RELEASES.md
@@ -8039,3 +8039,13 @@ Based on:
 - [typescript v4.7.0-rc.47] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.47] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.47 - packages/orq-rc
+
+## 2026-04-07 05:21:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.48] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.48 - packages/orq-rc